+++ agent/src/main/java/com/teamscale/jacoco/agent/testimpact/TestwiseCoverageAgent.java
@@ -28,6 +28,9 @@ public class TestwiseCoverageAgent extends AgentBase {
 			case TEAMSCALE_UPLOAD:
 				testEventHandler = new CoverageToTeamscaleStrategy(controller, options, reportGenerator);
 				break;
+			case DISK:
+				testEventHandler = new CoverageToDiskStrategy(controller, options, reportGenerator);
+				break;
 			case HTTP:
 				testEventHandler = new CoverageViaHttpStrategy(controller, options, reportGenerator);
 				break;