+++ common-system-test/src/main/java/com/teamscale/test/commons/TeamscaleMockServer.java
@@ -54,13 +54,18 @@ public class TeamscaleMockServer {
 	private final Service service;
 	private final List<String> impactedTests;
 
-	public TeamscaleMockServer(int port, String... impactedTests) throws IOException {
+	public TeamscaleMockServer(int port, boolean expectNoInteraction, String...impactedTests) throws IOException {
 		this.impactedTests = Arrays.asList(impactedTests);
 		service = Service.ignite();
 		service.port(port);
-		service.post("api/v5.9.0/projects/:projectName/external-analysis/session/auto-create/report",
-				this::handleReport);
-		service.put("api/v8.0.0/projects/:projectName/impacted-tests", this::handleImpactedTests);
+		if (expectNoInteraction) {
+			service.post("", (request, response) -> { throw new IllegalStateException("No POST to the mock server expected!"); });
+			service.put("", (request, response) -> { throw new IllegalStateException("No PUT to the mock server expected!"); });
+		} else {
+			service.post("api/v5.9.0/projects/:projectName/external-analysis/session/auto-create/report",
+					this::handleReport);
+			service.put("api/v8.0.0/projects/:projectName/impacted-tests", this::handleImpactedTests);
+		}
 		service.exception(Exception.class, (Exception exception, Request request, Response response) -> {
 			response.status(SC_INTERNAL_SERVER_ERROR);
 			response.body("Exception: " + exception.getMessage());