+++ system-tests/tia-maven/maven-dump-local-project/src/main/java/foo/SUT.java
@@ -1,0 +1,17 @@
+package foo;
+
+public class SUT {
+
+    public void bla() {
+        System.out.println("bla");
+    }
+
+    public void foo() {
+        System.out.println("foo");
+    }
+
+    public void goo() {
+        System.out.println("goo");
+    }
+
+}